Fix energy bar was not correctly rendered if more than 40 M FE/t was stored
(cherry picked from commit 2f4f8b8febaba10f6c04bd0cd29df48da9afb420)

+++ src/main/java/me/jddev0/ep/screen/base/IEnergyStorageConsumerIndicatorBarMenu.java
@@ -10,6 +10,6 @@ public interface IEnergyStorageConsumerIndicatorBarMenu extends IEnergyStorageMe
         int capacity = getCapacity();
 
         return (energyRequirement <= 0 || capacity == 0)?0:
-                (Math.min(energyRequirement, capacity - 1) *energyMeterHeight / capacity + 1);
+                (int)((long)Math.min(energyRequirement, capacity - 1) * energyMeterHeight / capacity + 1);
     }
 }
